dariah-demo-annotation/requirements.txt
kafka-python==2.0.2
requests==2.27.1
cloudevents==1.2.0